Splitting module
