+++ src/theory/quantifiers/quant_split.h
@@ -18,7 +18,7 @@
 #define CVC4__THEORY__QUANT_SPLIT_H
 
 #include "context/cdo.h"
-#include "theory/quantifiers/quant_util.h"
+#include "theory/quantifiers/quant_module.h"
 
 namespace CVC4 {
 namespace theory {
